+++ src/formats.rs
@@ -1,10 +1,10 @@
 use chrono::Local;
 use log::Record;
-use std::{io, thread};
+use std::thread;
 
 /// A logline-formatter that produces log lines like <br>
 /// ```INFO [my_prog::some_submodule] Task successfully read from conf.json```
-pub fn default_format(w: &mut io::Write, record: &Record) -> Result<(), io::Error> {
+pub fn default_format(w: &mut std::io::Write, record: &Record) -> Result<(), std::io::Error> {
     write!(
         w,
         "{} [{}] {}",
@@ -14,12 +14,32 @@ pub fn default_format(w: &mut io::Write, record: &Record) -> Result<(), io::Erro
     )
 }
 
+/// A colored version of the logline-formatter `default_format`
+/// that produces log lines like <br>
+/// <code><span style="color:red">ERROR</span> &#91;my_prog::some_submodule&#93; <span style="color:red">File not found</span></code>
+///
+/// Only available with feature `colors`.
+#[cfg(feature = "colors")]
+pub fn colored_default_format(
+    w: &mut std::io::Write,
+    record: &Record,
+) -> Result<(), std::io::Error> {
+    let level = record.level();
+    write!(
+        w,
+        "{} [{}] {}",
+        style(level, level),
+        record.module_path().unwrap_or("<unnamed>"),
+        style(level, record.args())
+    )
+}
+
 /// A logline-formatter that produces log lines like
 /// <br>
 /// ```[2016-01-13 15:25:01.640870 +01:00] INFO [src/foo/bar:26] Task successfully read from conf.json```
 /// <br>
 /// i.e. with timestamp and file location.
-pub fn opt_format(w: &mut io::Write, record: &Record) -> Result<(), io::Error> {
+pub fn opt_format(w: &mut std::io::Write, record: &Record) -> Result<(), std::io::Error> {
     write!(
         w,
         "[{}] {} [{}:{}] {}",
@@ -31,12 +51,29 @@ pub fn opt_format(w: &mut io::Write, record: &Record) -> Result<(), io::Error> {
     )
 }
 
+/// A colored version of the logline-formatter `opt_format`.
+///
+/// Only available with feature `colors`.
+#[cfg(feature = "colors")]
+pub fn colored_opt_format(w: &mut std::io::Write, record: &Record) -> Result<(), std::io::Error> {
+    let level = record.level();
+    write!(
+        w,
+        "[{}] {} [{}:{}] {}",
+        style(level, Local::now().format("%Y-%m-%d %H:%M:%S%.6f %:z")),
+        style(level, level),
+        record.file().unwrap_or("<unnamed>"),
+        record.line().unwrap_or(0),
+        style(level, &record.args())
+    )
+}
+
 /// A logline-formatter that produces log lines like
 /// <br>
 /// ```[2016-01-13 15:25:01.640870 +01:00] INFO [foo::bar] src/foo/bar.rs:26: Task successfully read from conf.json```
 /// <br>
 /// i.e. with timestamp, module path and file location.
-pub fn detailed_format(w: &mut io::Write, record: &Record) -> Result<(), io::Error> {
+pub fn detailed_format(w: &mut std::io::Write, record: &Record) -> Result<(), std::io::Error> {
     write!(
         w,
         "[{}] {} [{}] {}:{}: {}",
@@ -49,12 +86,33 @@ pub fn detailed_format(w: &mut io::Write, record: &Record) -> Result<(), io::Err
     )
 }
 
+/// A colored version of the logline-formatter `detailed_format`.
+///
+/// Only available with feature `colors`.
+#[cfg(feature = "colors")]
+pub fn colored_detailed_format(
+    w: &mut std::io::Write,
+    record: &Record,
+) -> Result<(), std::io::Error> {
+    let level = record.level();
+    write!(
+        w,
+        "[{}] {} [{}] {}:{}: {}",
+        style(level, Local::now().format("%Y-%m-%d %H:%M:%S%.6f %:z")),
+        style(level, record.level()),
+        record.module_path().unwrap_or("<unnamed>"),
+        record.file().unwrap_or("<unnamed>"),
+        record.line().unwrap_or(0),
+        style(level, &record.args())
+    )
+}
+
 /// A logline-formatter that produces log lines like
 /// <br>
 /// ```[2016-01-13 15:25:01.640870 +01:00] T[taskreader] INFO [src/foo/bar:26] Task successfully read from conf.json```
 /// <br>
 /// i.e. with timestamp, thread name and file location.
-pub fn with_thread(w: &mut io::Write, record: &Record) -> Result<(), io::Error> {
+pub fn with_thread(w: &mut std::io::Write, record: &Record) -> Result<(), std::io::Error> {
     write!(
         w,
         "[{}] T[{:?}] {} [{}:{}] {}",
@@ -65,4 +123,36 @@ pub fn with_thread(w: &mut io::Write, record: &Record) -> Result<(), io::Error> 
         record.line().unwrap_or(0),
         &record.args()
     )
+}
+
+/// A colored version of the logline-formatter `with_thread`.
+///
+/// Only available with feature `colors`.
+#[cfg(feature = "colors")]
+pub fn colored_with_thread(w: &mut std::io::Write, record: &Record) -> Result<(), std::io::Error> {
+    let level = record.level();
+    write!(
+        w,
+        "[{}] T[{:?}] {} [{}:{}] {}",
+        style(level, Local::now().format("%Y-%m-%d %H:%M:%S%.6f %:z")),
+        style(level, thread::current().name().unwrap_or("<unnamed>")),
+        style(level, level),
+        record.file().unwrap_or("<unnamed>"),
+        record.line().unwrap_or(0),
+        style(level, &record.args())
+    )
+}
+
+/// Helper function that is used in the provided colored format functions.
+///
+/// Only available with feature `colors`.
+#[cfg(feature = "colors")]
+pub fn style<T>(level: log::Level, item: T) -> yansi::Paint<T> {
+    match level {
+        log::Level::Error => yansi::Paint::fixed(196, item).bold(),
+        log::Level::Warn => yansi::Paint::fixed(208, item).bold(),
+        log::Level::Info => yansi::Paint::new(item),
+        log::Level::Debug => yansi::Paint::fixed(7, item),
+        log::Level::Trace => yansi::Paint::fixed(8, item),
+    }
 }