+++ src/logger.rs
@@ -56,7 +56,8 @@ pub struct Logger {
     parse_errs: Option<Vec<String>>,
     log_target: LogTarget,
     duplicate: Duplicate,
-    format: FormatFunction,
+    format_for_file: FormatFunction,
+    format_for_stderr: FormatFunction,
     flwb: FileLogWriterBuilder,
     other_writers: HashMap<String, Box<LogWriter>>,
 }
@@ -94,12 +95,18 @@ impl Logger {
     }
 
     fn from_spec_and_errs(spec: LogSpecification, parse_errs: Option<Vec<String>>) -> Logger {
+        #[cfg(feature = "colors")]
+        let default_format = formats::colored_default_format;
+        #[cfg(not(feature = "colors"))]
+        let default_format = formats::default_format;
+
         Logger {
             spec,
             parse_errs,
             log_target: LogTarget::StdErr,
             duplicate: Duplicate::None,
-            format: formats::default_format,
+            format_for_file: default_format,
+            format_for_stderr: default_format,
             flwb: FileLogWriter::builder(),
             other_writers: HashMap::<String, Box<LogWriter>>::new(),
         }
@@ -126,8 +133,8 @@ impl Logger {
     /// The factory methods `Logger::with_str()`, `Logger::with_env()`,
     /// and `Logger::with_env_or_str()`,
     /// parse a log specification String, and deduce from it a `LogSpecification` object.
-    /// Parsing errors are reported to stdout, but effectively ignored; in worst case, a
-    /// LogSpecification might be used that turns off logging completely!
+    /// Parsing errors are reported to stdout, but effectively ignored; in worst case, if the
+    /// log specification is completely broken, logging might be turned off completely!
     ///
     /// This method gives programmatic access to parse errors, if there were any.
     ///
@@ -182,15 +189,38 @@ impl Logger {
         self
     }
 
-    /// Makes the logger use the provided format function for the log entries,
-    /// rather than [formats::default_format](fn.default_format.html).
+    /// Makes the logger use the provided format function for all messages
+    /// that are written to files or to stderr.
     ///
-    /// You can either choose between some predefined variants,
-    /// ```default_format```, ```opt_format```, ```detailed_format```, ```with_thread```,
-    /// or you create and use your own format function
-    /// with the signature ```fn(&Record) -> String```.
+    /// You can either choose one of the provided log-line formatters,
+    /// or you create and use your own format function with the signature <br>
+    /// ```fn(&Record) -> String```.
+    ///
+    /// By default,
+    /// `default_format()` is used for the output to files and
+    /// `colored_default_format()` is used for the output to stderr.
+    ///
+    /// If the feature `colors` is switched off,
+    /// `default_format()` is used for all outputs.
     pub fn format(mut self, format: FormatFunction) -> Logger {
-        self.format = format;
+        self.format_for_file = format;
+        self.format_for_stderr = format;
+        self
+    }
+
+    /// Makes the logger use the provided format function for messages that are written to files.
+    ///
+    /// Regarding the default, see [Logger::format()](struct.Logger.html#method.format).
+    pub fn format_for_files(mut self, format: FormatFunction) -> Logger {
+        self.format_for_file = format;
+        self
+    }
+
+    /// Makes the logger use the provided format function for messages that are written to stderr.
+    ///
+    /// Regarding the default, see [Logger::format()](struct.Logger.html#method.format).
+    pub fn format_for_stderr(mut self, format: FormatFunction) -> Logger {
+        self.format_for_stderr = format;
         self
     }
 
@@ -220,7 +250,8 @@ impl Logger {
         self
     }
 
-    /// Use rotation to prevent indefinite growth of log files.
+    /// Prevent indefinite growth of the log file by applying file rotation
+    /// and a clean-up strategy for older log files.
     ///
     /// By default, the log file is fixed while your program is running and will grow indefinitely.
     /// With this option being used, when the log file reaches the specified criterion,
@@ -246,7 +277,12 @@ impl Logger {
     /// my_prog_rCURRENT.log
     /// ```
     ///
-    /// The cleanup parameter allows defining the strategy for dealing with older files.
+    /// ## Parameters
+    ///
+    /// `rotate_over_size` is given in bytes, e.g. `10_000_000` will rotate
+    /// files once they reach a size of 10 MiB.
+    ///     
+    /// `cleanup` defines the strategy for dealing with older files.
     /// See [Cleanup](enum.Cleanup.html) for details.
     pub fn rotate(mut self, criterion: Criterion, naming: Naming, cleanup: Cleanup) -> Logger {
         self.flwb = self.flwb.rotate(criterion, naming, cleanup);
@@ -384,7 +420,7 @@ impl Logger {
     }
 }
 
-/// Start logging.
+/// Finally, start logging, optionally with a spec-file.
 impl Logger {
     /// Consumes the Logger object and initializes `flexi_logger`.
     ///
@@ -398,11 +434,15 @@ impl Logger {
 
         let primary_writer = Arc::new(match self.log_target {
             LogTarget::File => {
-                self.flwb = self.flwb.format(self.format);
-                PrimaryWriter::file(self.duplicate, self.flwb.try_build()?)
+                self.flwb = self.flwb.format(self.format_for_file);
+                PrimaryWriter::file(
+                    self.duplicate,
+                    self.format_for_stderr,
+                    self.flwb.try_build()?,
+                )
             }
-            LogTarget::StdErr => PrimaryWriter::stderr(self.format),
-            LogTarget::DevNull => PrimaryWriter::black_hole(self.duplicate, self.format),
+            LogTarget::StdErr => PrimaryWriter::stderr(self.format_for_stderr),
+            LogTarget::DevNull => PrimaryWriter::black_hole(self.duplicate, self.format_for_stderr),
         });
 
         let flexi_logger = FlexiLogger::new(