+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.12.0] - 2019-unknown
+## [0.12.0] - 2019-05-24
 
 Revise handling of record.metadata().target() versus record.module_path().
 
-Incompatible  API modification: Logger.rotate() takes now three parameters.
+Incompatible API modification: Logger.rotate() takes now three parameters.
+
+Suppport different formatting for stderr and files.
+
+Add feature `colors` (see `README.md` for details).
+
+Remove the deprecated `Logger::start_reconfigurable()` and `Logger::rotate_over_size()`.
 
 ## [0.11.5] - 2019-05-15